requirements.txt
# Deep Learning Framework
torch>=1.12.0
torchvision>=0.13.0

# Reinforcement Learning
stable-baselines3>=2.0.0
gymnasium>=0.29.0
shimmy>=1.0.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0

# DEM Processing
rasterio>=1.3.0
pyproj>=3.4.0

# Visualization
matplotlib>=3.5.0
plotly>=5.10.0

# Configuration
pyyaml>=6.0
argparse

# Logging
tensorboard>=2.10.0
tqdm>=4.64.0

# Utilities
pandas>=1.4.0
